logout functionality

+++ base/views.py
@@ -1,19 +1,24 @@
 from django.shortcuts import render, redirect
 from django.db.models import Q
+from django.http import HttpResponse
+from django.contrib.auth.decorators import login_required
 from .models import Room, Topic
 from .forms import RoomForm
 from django.contrib.auth.models import User
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.forms import UserCreationForm
 
 #create a rooms
 rooms = []
 
 # views
-
 def loginPage(request):
+    page = 'login'
+    if request.user.is_authenticated:
+        return redirect('home')
     if request.method == 'POST':
-        username = request.POST.get('username')
+        username = request.POST.get('username').lower()
         password = request.POST.get('password')
 
         try:
@@ -28,12 +33,33 @@ def loginPage(request):
             messages.error(request, 'Username or password error')
 
 
-    context = {}
+    context = {
+        'page' : page
+    }
     return render(request, 'base/login_registration.html', context)
 
 def Logoutuser(request):
     logout(request)
     return redirect('home')
+
+
+def registerPage(request):
+    form = UserCreationForm()
+
+    if request.method == 'POST':
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occurred while registering")
+    context = {
+        'form' : form,
+    }
+    return render(request, 'base/login_registration.html', context)
 
 def home(request):
     q = request.GET.get('q') if request.GET.get('q') != None else ''
@@ -53,7 +79,7 @@ def room(request, pk):
     context = {'room' : room}
     return render(request, 'base/room.html', context)
 
-
+@login_required(login_url = 'login')
 def createRoom(request):
     form = RoomForm()
     if request.method == 'POST':
@@ -64,10 +90,13 @@ def createRoom(request):
     context = {'form' : form}
     return render(request, 'base/room_form.html', context)
 
-
+@login_required(login_url = 'login')
 def  updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed")
 
     if request.method == 'POST':
         form = RoomForm(request.POST, instance=room)
@@ -78,8 +107,11 @@ def  updateRoom(request, pk):
     context = {'form': form}
     return render(request, 'base/room_form.html', context)
 
+@login_required(login_url = 'login')
 def deleteRoom(request, pk):
-    room = Room.objects.get(id=pk)  
+    room = Room.objects.get(id=pk) 
+    if request.user != room.host:
+        return HttpResponse("You are not allowed") 
     if request.method == 'POST':
         room.delete()
         return redirect('home')
